Implement Distribution (#7)
* wip

* Fix errors & mistakes

* on_initialize seems to have to be top

* why error

* test commout

* type

* reenable oninit

* type

* Finally I got the huge shit

+++ runtime/src/types.rs
@@ -6,7 +6,7 @@ use sp_core::{Blake2Hasher, Hasher, H256};
 pub type AccountId = H256;
 pub type Signature = Vec<u8>;
 pub type Nonce = u64;
-pub type Balance = u64;
+pub type Balance = i128;
 
 /// The struct of individual account
 #[derive(Encode, Decode, Default, Clone, PartialEq)]
